Remove old IslandCreationPanel and move to the new one.
