+++ src/main/java/world/bentobox/bentobox/api/commands/island/IslandResetCommand.java
@@ -16,7 +16,7 @@ import world.bentobox.bentobox.database.objects.Island;
 import world.bentobox.bentobox.managers.BlueprintsManager;
 import world.bentobox.bentobox.managers.island.NewIsland;
 import world.bentobox.bentobox.managers.island.NewIsland.Builder;
-import world.bentobox.bentobox.panels.IslandCreationPanel;
+import world.bentobox.bentobox.panels.customizable.IslandCreationPanel;
 import world.bentobox.bentobox.util.Util;
 
 /**
